Split up TestTempTableName tests
